doty: fix shebang and add execution bit
DayOfTheYear.py had a shebang that did not support running from within a python
virtualenv: it would always invoke the system's python interpreter.

Moreover, the presence of a shebang indicates that the script supports being
called directly from the shell. In that case we also need to set the execution
bit.